+++ runtime/datatypes/tag.ts
@@ -1,5 +1,6 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
+import {insert} from "./string-ref";
 
 /* Actions */
 export function $$form(
@@ -22,4 +23,17 @@ export function $$mold(
 	_: RedActions.MoldOptions = {}
 ): boolean {
 	return $$form(ctx, tag, buffer, _.part);
+}
+
+// ...
+
+export function $$insert(
+	ctx:   Red.Context,
+	tag:   Red.RawTag,
+	value: Red.AnyType,
+	_: RedActions.InsertOptions = {}
+): Red.RawTag {
+	tag.index += insert(ctx, tag.tag, tag.index - 1, value, _);
+	
+	return tag;
 }