+++ runtime/datatypes/url.ts
@@ -1,5 +1,6 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
+import {insert} from "./string-ref";
 
 /* Actions */
 export function $$form(
@@ -20,4 +21,17 @@ export function $$mold(
 	_: RedActions.MoldOptions = {}
 ): boolean {
 	return $$form(ctx, url, buffer, _.part);
+}
+
+// ...
+
+export function $$insert(
+	ctx:   Red.Context,
+	url:   Red.RawUrl,
+	value: Red.AnyType,
+	_: RedActions.InsertOptions = {}
+): Red.RawUrl {
+	url.index += insert(ctx, url.url, url.index - 1, value, _, encodeURI);
+	
+	return url;
 }